etc/database/screen-capture-annotate/requirements.txt
selenium==4.23.1
requests==2.32.3
python-dotenv==1.0.1
yapf==0.40.2